requirements.txt
absl-py==2.0.0
aiobotocore==2.9.0
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.2.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
boto3==1.34.16
botocore==1.34.16
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
flatbuffers==23.5.26
frozenlist==1.4.1
fs==2.4.16
fs-s3fs==1.1.1
fsspec==2023.12.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.13.1
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.4.2
Pillow==9.1.1
protobuf==4.25.1
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
s3fs==2023.12.2
s3transfer==0.10.0
scikit-learn==1.3.0
scipy==1.7.3
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.5.0
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
